# frontend/lib/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "allwin API",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/redeem": {
   "post": {
    "tags": [
     "member"
    ],
    "summary": "Redeem",
    "operationId": "redeem_api_v1_redeem_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RedeemBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RedeemResponse"
        }
       }
      }
     },
     "400": {
      "description": "Bad Request",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "401": {
      "description": "Unauthorized",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "403": {
      "description": "Forbidden",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "404": {
      "description": "Not Found",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "422": {
      "description": "Unprocessable Content",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/favorites": {
   "get": {
    "tags": [
     "member"
    ],
    "summary": "List Favorites",
    "operationId": "list_favorites_api_v1_favorites_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/FavoritesResponse"
        }
       }
      }
     },
     "400": {
      "description": "Bad Request",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "401": {
      "description": "Unauthorized",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "403": {
      "description": "Forbidden",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "404": {
      "description": "Not Found",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "422": {
      "description": "Unprocessable Content",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "member"
    ],
    "summary": "Add Favorite",
    "operationId": "add_favorite_api_v1_favorites_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/FavoriteBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/OkDTO"
        }
       }
      }
     },
     "400": {
      "description": "Bad Request",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "401": {
      "description": "Unauthorized",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "403": {
      "description": "Forbidden",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "404": {
      "description": "Not Found",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "422": {
      "description": "Unprocessable Content",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/favorites/{match_id}": {
   "delete": {
    "tags": [
     "member"
    ],
    "summary": "Remove Favorite",
    "operationId": "remove_favorite_api_v1_favorites__match_id__delete",
    "parameters": [
     {
      "name": "match_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Match Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/OkDTO"
        }
       }
      }
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Bad Request"
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Unauthorized"
     },
     "403": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Forbidden"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Not Found"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      },
      "description": "Unprocessable Content"
     }
    }
   }
  },
  "/api/v1/account": {
   "get": {
    "tags": [
     "member"
    ],
    "summary": "Account",
    "operationId": "account_api_v1_account_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AccountResponse"
        }
       }
      }
     },
     "400": {
      "description": "Bad Request",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "401": {
      "description": "Unauthorized",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "403": {
      "description": "Forbidden",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "404": {
      "description": "Not Found",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "422": {
      "description": "Unprocessable Content",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/account/sessions/revoke": {
   "post": {
    "tags": [
     "member"
    ],
    "summary": "Revoke Other Session",
    "operationId": "revoke_other_session_api_v1_account_sessions_revoke_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RevokeSessionBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/OkDTO"
        }
       }
      }
     },
     "400": {
      "description": "Bad Request",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "401": {
      "description": "Unauthorized",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "403": {
      "description": "Forbidden",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "404": {
      "description": "Not Found",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     },
     "422": {
      "description": "Unprocessable Content",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiErrorDTO"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "AccountIdentity": {
    "properties": {
     "provider": {
      "type": "string",
      "title": "Provider"
     },
     "provider_app_id": {
      "type": "string",
      "title": "Provider App Id"
     },
     "created_at": {
      "type": "string",
      "title": "Created At"
     },
     "last_used_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Used At"
     }
    },
    "type": "object",
    "required": [
     "provider",
     "provider_app_id",
     "created_at"
    ],
    "title": "AccountIdentity"
   },
   "AccountRecovery": {
    "properties": {
     "available": {
      "type": "boolean",
      "title": "Available"
     },
     "note": {
      "type": "string",
      "title": "Note"
     }
    },
    "type": "object",
    "required": [
     "available",
     "note"
    ],
    "title": "AccountRecovery"
   },
   "AccountResponse": {
    "properties": {
     "user": {
      "$ref": "#/components/schemas/MeUser"
     },
     "plan": {
      "type": "string",
      "title": "Plan"
     },
     "entitlements": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Entitlements"
     },
     "identities": {
      "items": {
       "$ref": "#/components/schemas/AccountIdentity"
      },
      "type": "array",
      "title": "Identities"
     },
     "subscriptions": {
      "items": {
       "$ref": "#/components/schemas/AccountSubscription"
      },
      "type": "array",
      "title": "Subscriptions"
     },
     "sessions": {
      "items": {
       "$ref": "#/components/schemas/AccountSession"
      },
      "type": "array",
      "title": "Sessions"
     },
     "recovery": {
      "$ref": "#/components/schemas/AccountRecovery"
     }
    },
    "type": "object",
    "required": [
     "user",
     "plan",
     "entitlements",
     "identities",
     "subscriptions",
     "sessions",
     "recovery"
    ],
    "title": "AccountResponse"
   },
   "AccountSession": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "created_at": {
      "type": "string",
      "title": "Created At"
     },
     "last_seen_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Seen At"
     },
     "expires_at": {
      "type": "string",
      "title": "Expires At"
     },
     "user_agent": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "User Agent"
     },
     "is_current": {
      "type": "integer",
      "title": "Is Current"
     }
    },
    "type": "object",
    "required": [
     "id",
     "created_at",
     "expires_at",
     "is_current"
    ],
    "title": "AccountSession"
   },
   "AccountSubscription": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "plan_id": {
      "type": "string",
      "title": "Plan Id"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "starts_at": {
      "type": "string",
      "title": "Starts At"
     },
     "ends_at": {
      "type": "string",
      "title": "Ends At"
     },
     "source": {
      "type": "string",
      "title": "Source"
     },
     "created_at": {
      "type": "string",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "plan_id",
     "status",
     "starts_at",
     "ends_at",
     "source",
     "created_at"
    ],
    "title": "AccountSubscription"
   },
...
   "ApiErrorDTO": {
    "properties": {
     "code": {
      "type": "string",
      "title": "Code"
     },
     "message": {
      "type": "string",
      "title": "Message"
     },
     "details": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Details"
     }
    },
    "type": "object",
    "required": [
     "code",
     "message",
     "details"
    ],
    "title": "ApiErrorDTO"
   },
...
   "FavoriteBody": {
    "properties": {
     "match_id": {
      "type": "integer",
      "title": "Match Id"
     }
    },
    "type": "object",
    "required": [
     "match_id"
    ],
    "title": "FavoriteBody"
   },
   "FavoriteItem": {
    "properties": {
     "match_id": {
      "type": "integer",
      "title": "Match Id"
     },
     "created_at": {
      "type": "string",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "match_id",
     "created_at"
    ],
    "title": "FavoriteItem"
   },
   "FavoritesResponse": {
    "properties": {
     "favorites": {
      "items": {
       "$ref": "#/components/schemas/FavoriteItem"
      },
      "type": "array",
      "title": "Favorites"
     }
    },
    "type": "object",
    "required": [
     "favorites"
    ],
    "title": "FavoritesResponse"
   },
...
   "MeUser": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "display_name": {
      "type": "string",
      "title": "Display Name"
     },
     "role": {
      "type": "string",
      "title": "Role"
     }
    },
    "type": "object",
    "required": [
     "id",
     "display_name",
     "role"
    ],
    "title": "MeUser"
   },
...
   "OkDTO": {
    "properties": {
     "status": {
      "type": "string",
      "const": "ok",
      "title": "Status"
     }
    },
    "type": "object",
    "required": [
     "status"
    ],
    "title": "OkDTO"
   },
...
   "RedeemBody": {
    "properties": {
     "code": {
      "type": "string",
      "maxLength": 32,
      "minLength": 6,
      "title": "Code"
     }
    },
    "type": "object",
    "required": [
     "code"
    ],
    "title": "RedeemBody"
   },
...
   "RedeemResponse": {
    "properties": {
     "status": {
      "type": "string",
      "const": "ok",
      "title": "Status"
     },
     "subscription_id": {
      "type": "string",
      "title": "Subscription Id"
     },
     "plan_id": {
      "type": "string",
      "title": "Plan Id"
     },
     "starts_at": {
      "type": "string",
      "title": "Starts At"
     },
     "ends_at": {
      "type": "string",
      "title": "Ends At"
     }
    },
    "type": "object",
    "required": [
     "status",
     "subscription_id",
     "plan_id",
     "starts_at",
     "ends_at"
    ],
    "title": "RedeemResponse"
   },
...
   "RevokeSessionBody": {
    "properties": {
     "session_id": {
      "type": "string",
      "title": "Session Id"
     }
    },
    "type": "object",
    "required": [
     "session_id"
    ],
    "title": "RevokeSessionBody"
   },
...
  }
 }
}